Default background colors for some message types
For example, if "Warning: blah blah blah" is entered as the text, the background will be set to yellow.

This functionality is NOT case-sensitive, so "wArnIng" will also be yellow. It also ignores punctuation, so "Warning:" is the same as "Warning"

If a custom colour is passed as an argument in styles, then it will be used instead of the default colour for the type.

If no colour was found for the message, no colour will be selected.